+++ albedo-plugins/albedo-data-mybatis/src/main/java/com/albedo/java/common/persistence/service/impl/TreeServiceImpl.java
@@ -100,7 +100,7 @@ public class TreeServiceImpl<Repository extends TreeRepository<T>, T extends Tre
 	}
 
 	@Override
-	public Integer countByParentId(String parentId) {
+	public Long countByParentId(String parentId) {
 		return repository.selectCount(Wrappers.<T>query().eq(TreeEntity.F_SQL_PARENT_ID, parentId));
 	}
 
@@ -126,7 +126,7 @@ public class TreeServiceImpl<Repository extends TreeRepository<T>, T extends Tre
 		}
 
 		if (ObjectUtil.isNotEmpty(entityDto.getId())) {
-			Integer count = countByParentId(entityDto.getId());
+			Long count = countByParentId(entityDto.getId());
 			entityDto.setLeaf(count == null || count == 0);
 		} else {
 			entityDto.setLeaf(true);
